fix: address Sourcery review on PR #100
- Move sanitize_host from config_flow to helper so migration.py no
  longer needs to import from config_flow, avoiding tight coupling
  and future circular-import risk.
- Normalize both sides of the host comparison in
  TrueNASConfigFlow._find_legacy_config via sanitize_host, matching
  migration._find_legacy_entry so a legacy host stored in a
  different case (e.g. NAS.local) is found consistently by both.
