+++ custom_components/truenas_ce/migration.py
@@ -34,7 +34,6 @@ from homeassistant.helpers import entity_registry as er
 from homeassistant.helpers.storage import Store
 from homeassistant.util import dt as dt_util
 
-from .config_flow import _sanitize_host
 from .const import (
     DOMAIN,
     LEGACY_DOMAIN,
@@ -44,6 +43,7 @@ from .const import (
     MIGRATION_LEGACY_ENTRY_ID,
     MIGRATION_RECORDS,
 )
+from .helper import sanitize_host
 
 _LOGGER = getLogger(__name__)
 
@@ -164,13 +164,13 @@ def _find_legacy_entry(
 ) -> ConfigEntry | None:
     """Find the old ``truenas`` config entry matching this one (by host).
 
-    Both sides are run through ``_sanitize_host`` before comparing, so a
+    Both sides are run through ``sanitize_host`` before comparing, so a
     legacy host stored in a different case (e.g. ``NAS.local``) still matches.
     """
     candidates = hass.config_entries.async_entries(LEGACY_DOMAIN)
-    host = _sanitize_host(config_entry.data.get(CONF_HOST, ""))
+    host = sanitize_host(config_entry.data.get(CONF_HOST, ""))
     for entry in candidates:
-        if _sanitize_host(entry.data.get(CONF_HOST, "")) == host:
+        if sanitize_host(entry.data.get(CONF_HOST, "")) == host:
             return entry
     # Single legacy entry with a differing host (e.g. host was normalized): adopt it.
     return candidates[0] if len(candidates) == 1 else None
